[iOS] Fix textfield selection in Settings Autofill edit
When the user tap on a cell of the Edit screen of an Address, the
associated text field should be focused.
This wasn't the case because the selection of the cell isn't registered
if the cell cannot be edited.

Bug: 989486
Change-Id: I5f193c59260151aa811cd15b92ebe2959476db18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728563
Reviewed-by: Yi Su <mrsuyi@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#682696}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 599c552d0080f557afcd8be28112318a0df19c34