[iOS] Adding ChromeIdentityService::GetCachedHostedDomainForIdentity()
When sign-in with a new gmail account, the Chrome identity was
initialized with no hosted domain since the SSO profile was not fetched
yet.
The property was never updated later. Even when the AuthenticationFlow
was fetching hosted domain using the AuthenticationFlowPerformer.
During the sign-in process

To fix the issue, -ChromeIdentity.hostedDomain is remove. ChromeIdentity
stays just a proxy for SSOIdentity.
ChromeIdentityService::GetCachedHostedDomainForIdentity() is added to
get the hosted domain.

Bug introduced with:
  + crrev.com/c/1687254
  + crrev.com/i/1434627
  + crrev.com/c/1687254

Related to:
  => crrev.com/c/1732084 (adding new API) <=
  + crrev.com/i/1560989 (adding new implementation)
  + crrev.com/c/1732085 (switching implementation)
  + crrev.com/i/1560990 (old implementation cleanup)
  + crrev.com/c/1732101 (cleanup API)

Bug: 987380
Change-Id: I2dab7a36927530ca400bdf6e99d645ff325159c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732084
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#683588}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23d73b9fe2a8fdd5e27f7830b1f8461999b9086d