[iOS] Add separator to omnibox popup on iPad
After changing the toolbar background to be white, the omnibox popup
(which shares a background color) blends in with the NTP and any
website with a white background. This adds a separator to the bottom
of the popup to prevent this.

Bug: 976662
Change-Id: I31cb40a787a0859a02863ab8d300e43bf945339f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1741909
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Auto-Submit: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#684812}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ed1ae0a650e66eb8145483722ce16026b224fb58