[ios] Toggle badge view constraints in response to FullScreen progress
When full screen is enabled, the badge view should hug the right side
so that the incognito badge in dark mode can be right next to the
LocationBarContainerView. When full screen progress reaches down to 0.1,
the constraints will be changed to reflect that.

Bug: 976901
Change-Id: I3e258ba6280ebf060c7b306b6be80ae98ea1a6ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728711
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#684971}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe093d24f81085083cef4def1de8ddaafddc1fef