[leveldb_proto] Moved ProtoDatabaseProvider to StoragePartition
leveldb_proto is a component that allows storing proto objects in a
leveldb database. A client can get a leveldb_proto database by calling a
static method.

This simple API made it easy for developers to integrate leveldb_proto
in their components, this resulted in a large amount of clients, with
each one using a unique leveldb instance. Having multiple db instances
increases memory consumption, as each database caches frequently used
contents.

There has been an effort to move leveldb_proto databases from multiple
instances to a shared instance, to decrease memory consumption and
simplify configuration. As part of this effort there have been many
changes to the database retrieval API.

Our first approach was to use a ProtoDatabaseProvider instead of a
static method. This provider owns the shared instance, so there should
be one provider per Profile. To do this we created it with a
ProtoDatabaseProviderFactory, which was a
BrowserContextKeyedServiceFactory.

Some clients need to initialize their database in reduced mode,
so they don't have access to a BrowserContext/Profile. Instead
they have a SimpleFactoryKey, which is moved to a Profile once
a full startup occurs. Our second approach was to change our
ProtoDatabaseProviderFactory to a SimpleKeyedServiceFactory.

This worked for most clients, but some of them need to initialize
their database from BrowserContext itself, which can't access
SimpleFactoryKey because it lives in //components/keyed_service.

This CL removes ProtoDatabaseProviderFactory, ProtoDatabaseProvider
is now created and owned by StoragePartition for all cases except
for reduced mode. In reduced mode the DB provider is created by
StartupData, and then it transfers it to StoragePartition when
initializing in full mode.

In addition we also removed IOS's ProtoDatabaseProviderFactory,
ProtoDatabaseProvider is now created and owned by BrowserState.

Bug: 943354
Change-Id: I7cdeca4bff33a910969e21ed3d879938cf43842c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1656712
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: ssid <ssid@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685259}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b9d633476253c3a4dbb23d40f35a2c8f27f9a2e