[iOS] Fix jumping keyboard accessory view buttons on iPads.
There's a bug in iOS 13 that causes input view's items to jump when
they are defined as custom views. This CL replaces them with equivalent
button items with an image. This means a few things:
- Items don't jump anymore
- The code to create the buttons is no longer shared between the iPad
and the iPhone implementations
- The button shadows are no longer custom on iPad

This seems like a good enough fix for now. If this is not fixed in a
future version of iOS 13, we should probably create custom button icons
with a baked in shadow and consider a way to share the a11y strings and
icons between iPhone and iPad.

the voice search and the QR code scanner buttons don't jump

Bug: 984905
Test: On an iOS 13 iPad, type anything in the omnibox. Observe that
Change-Id: Ia19f1a3877c58e03b2ea5555c53ecfdfef9425b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742139
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685575}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7c13ca9b72308299ff513ad39e81ed3544c932d