[ios] Add buttonFactory as init param for BadgeViewController
This will make it possible to use it in Showcase. Both the buttonfactory
and the stackView of the BadgeViewController have to be initialized in
init in order for addBadge: to successfully add badges from a
Showcase coordinator.

Bug: 976901
Change-Id: I2a98150cbe64e3b407295224514da6220d393ec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1755050
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#687395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cdd0e2dc314bd238136b29f6c237c03051f02a54