[iOS][DarkMode] Tint the accessories with blue_color
Sets the tint color of the cell so when the accessory view is created
it is tinted with the right color.
Sets the tint color of the table view to blue so the index uses it.

Screenshots:
https://drive.google.com/file/d/1xfOJVT6ZqqOiSb_pEX_uWqeLeEcxpc2F/view?usp=sharing
https://drive.google.com/file/d/1-HwGeIh1d-vk3fSDfsLVxWTm0y64kRfG/view?usp=sharing

Bug: 976675
Change-Id: I3f1b4990060a1db29eb119c271d3402c8e1706b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749010
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#688524}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4be0712c3660f7d722d395451b34115da10f0317