Replace site information's gesture recognizer with a button.
It seems like the gesture recognizer was used to have a "button with a
hole" on the site info's background. This worked, but was not accessible
to voiceover. Probably a button was impossible to use in the past.

This CL replaces the gesture recognizer with a regular button, and the
"hole" is made by setting userInteractionEnabled to YES on the actual
content.
Also adds support for the a11y escape gesture.

Bug: 979789
Change-Id: Ib29318232ef216c1134f16e69c1e0e39cf5a099a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763417
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#688968}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be41285738c746c999ce3ad9e04da8a9181eb6ea