CookieStore callback: Change signature to accommodate inclusion statuses
This changes the CookieStore's GetCookieListCallback from taking a
CookieList (for included cookies) and a CookieStatusList (for excluded
cookies), to taking a CookieStatusList for both. Previously, the
statuses of included cookies could be assumed to be INCLUDE, so they
did not need to be passed to the callback. The next change adds more
nuanced information to CookieInclusionStatus such that statuses must
be explicitly passed even for included cookies, hence a CookieStatusList
must be provided.

This also makes GetAllCookiesCallback take a CookieList only, since the
second CookieStatusList parameter (for excluded cookies, of which there
were none, since we are getting all cookies) was unused.

Bug: 993843
Change-Id: I4c6bedc0c161f863b9a1755bb93e615f2793202e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1758437
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#689162}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f068a76f1819af40b5b5077fbcadc2381c1671d7