[iOS] Fix color of ReadingList empty string image
This CL changes the color of the image in the empty string of the
ReadingList. In iOS 13 the tint color of the screen isn't used for the
text attachment. A new API is created to have a tint image directly.

Bug: 982262
Change-Id: I97e6e1a336f3752e2c09ec0b6ae397beaee356fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695554
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#689872}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b043d13f3e769600452f63456f12f118b41d1f64