[iOS] Fixing notification from SigninPromoViewMediator to its consumer
SigninPromoViewMediator should not send update notification to its consumer
when the sign-in is in progress.
This is to avoid updating the sign-in promo from cold state to warm state
while the sign-in is not over.

Bug: 986746
Change-Id: I4f4386034549394eb556e6e48254d97dc4590656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1718370
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#691117}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07373d4c2903bf4be0896af7780ace9272144a58