[iOS] Fix JS exceptions thrown for iframe elements without src attribute
A JavaScript URL can not be created with an empty string. Attempting to
do so will throw an exception, so check if the strings are non empty
first.

Additionally, postMessage can not be called with an empty string for the
targetOrigin parameter. If the iframe element does not have a src
attribute, use “*” instead to forward the context menu message to a
child frame.

Bug: 911899
Change-Id: I866bec11fa48c3a0ef5b95a6a628a64384d9da59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1779423
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#693050}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 40ab560f6ad584be07ff02f8d355d445ffd02731