[iOS] Add gn-argument-gated plist/storyboard for multi-window.
This CL set up the iOS app configuration needed for multi-window. This
includes (a) a bunch of specific Info.plist additions, and (b) a
storyboard file to be used as a default by each scene.

These files are added if the gn argument ios_enable_multi_window is
set to 'true'. By default it is 'false'. Note that enabling this arg
will, in Chromium's current state, cause the app to launch with an empty
window and do nothing (because there's no class to load as the scene
delegate). Follow-up CLs will fill that gap, but they need the plist and
storyboard in place first.

Change-Id: I61e153cd2d5d5df12bbd6347151f864e44a2d4ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1782815
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#693137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9d47515368fbf42aa6446a26d373eb6fd724f6a