[iOS][Dark Mode] Add dark mode colors to side swipe-to-navigate
There are two major changes to the design here. First, the arrow icon
is now always the same color. It no longer changes color when the
circle is visible. Second, the image for the border between the
side_swipe view and the webpage has been removed. Now, it's always just
a plain UIView with a background color. This UIView looks the same as
the shadow between the toolbar and the web content (or an infobar and
the web content).

Bug: 998441
Change-Id: I0b96def297b7ab7688ea978ddc09d5a38be83c66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1768593
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#693184}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 483c4859c0a4b9eba8b9a2cb8414b5ca0b8bb31f