[SlimNav] Use targetURL when reloading restore session error pages.
When ones launches while offline and then reloads, the current URL
is a restore session URL but the target URL for the reload is not.
So when we call AddPendingItem in registerLoadRequestForURL,
WKBasedNavigationManagerImpl::AddPendingItem adds a new |pending_item_|
instead of setting |pending_item_index_| to the current index.

The new item has a new ErrorRetryStateMachine, in state kNewRequest,
so when the provisional navigation fails,
ErrorRetryStateMachine::DidFailProvisionalNavigation returns kLoadPlaceholder,
and loading the placeholder URL will clobber the forward list.

Per ajuma's advice, this CL changes WKBasedNavigationManagerImpl::AddPendingItem
to reuse the current item if its URL is a restore session URL for the target url.

Bug: 979589
Change-Id: I6680f14f90bc5b092cec2a3d57fd67352d169e0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1778981
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#693246}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c546ce63dea7d04254531e6efcfaee716c884953