[iOS][Dark Mode] Change tab switcher background color to black
This CL changes the tab switcher background color by adding a new
local color to the tab_grid resources. It also defines a constant
in tab_grid_constants to hold the name of this file and make sure that
anything that uses the color depends on the actual colorset.

It also deletes all the now-unused color constants. The only ones left
are those that are used for styling the recent tabs view on iOS 12
(where dark mode is not supported).

Bug: 998415
Change-Id: I9c7ac49360e36ad8681492ef7722436d9b5aa308
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774278
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#693324}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 782398d4149413ec9a0085ba612c2fcf8dcfb5c4