Add PasswordForm::from_store flag
This tags a PasswordForm as coming from either the profile-scoped
store or the account-scoped store (or "not set", as the default value
or for parsed forms). It will be used to differentiate these
credentials in the filling UI and in settings.

Bug: 998455
Change-Id: Idb49c7f0a5e659eb0e070c489adedd20bf192c73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749187
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Auto-Submit: Marc Treib <treib@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#694233}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6dbd036d25e6c9ecdd0168d97b206c5ddfc5823