Make visible_url_egtest.mm compatible with EG2
This adds the file as an EG2 target, and updates #includes.

This also adds a displayTitleForURL method to ChromeEarlGrey and
ChromeEarlGreyAppInterface. This provides the functionality of
web::GetDisplayTitleForUrl. The logic of parsing a URL in this manner
cannot be moved to the test process, since it depends on app-process-specific
logic; for example, it depends on  ChromeWebClient::AddAdditionalSchemes for
registering the "chrome:" scheme in order to correctly parse chrome:// URLs.

Bug: 987646
Change-Id: I2ef13336dae4dad1b9a774713c35f8c08e680233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1790602
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#694789}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ccf37713bf12daf2f9db52fa9b9d792396c2c466