Stop disabling WebUsage before clearing browsing data for iOS 13
This should fix the problem of cookies not clearing consistently
on iOS 13.
More details are on the crbug comments.

Also Change onViewUpdated on web_frames_manager_impl to remove all
messages instead of looking for specific messages. This is needed
as removing specific messages DCHECKs if the message is not registered.
and during the Purge which happens in clear browsing data the message
router is destroyed so there will be no messages in the registered with
the webview.

TEST= I tested manually on ipad, iphone for both iOS 12 & 13 on devices
, a test on the simulator is meaningless as the simulator doesn't use
the same cookie storage that the device use. There exist unit tests but
they will not fail in the case that i'm fixing because in these tests
we set cookies manually and this current case happen when cookies are set
by the web view. I'll discuss further creating EG tests to test the failed
scenario using a browsing test page that can set/get cookies.

Bug: 997172, 956516
Change-Id: I3befee1bf38ac1d3d86c33e50bbbbd4fb9e75ec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1791629
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#695709}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d03761396d4e85f8d42ffa571377acff9ba3b6f8